Apply PR suggestions
